# openapi.yaml
openapi: 3.0.3
info:
  title: Fixed-Slot Appointment Booking API
  description: >
    Production-grade RESTful API for fixed-slot appointment booking with real-time updates via Socket.IO,
    strict concurrency conflict prevention using PostgreSQL row-level locks and database constraints,
    and seed data management.
  version: 1.0.0
servers:
  - url: http://localhost:4000
    description: Local Development / Docker Server
paths:
...
  /slots:
    get:
      summary: List fixed appointment slots
      description: Returns a list of fixed appointment slots. Can filter for available slots only using the `available` query parameter.
      parameters:
        - name: available
          in: query
          required: false
          description: If true, returns only available (unbooked) slots.
          schema:
            type: boolean
            example: true
      responses:
        '200':
          description: List of appointment slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Slot'
...
  /bookings:
    post:
      summary: Create booking for a fixed slot
      description: >
        Reserves a fixed slot. Uses PostgreSQL row-level locking (`SELECT ... FOR UPDATE`)
        within an ACID transaction and partial unique database constraints to guarantee
        that concurrent requests for the same slot never result in double booking.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookingRequest'
            example:
              slotId: 4a2f8b50-3a1b-4f9e-9d22-123456789abc
              clientName: Ahmed Al-Mansoor
              clientEmail: ahmed@example.com
      responses:
        '201':
          description: Booking successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '400':
          description: Validation error in request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '404':
          description: Slot does not exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict - Slot is already booked (or claimed by a concurrent request)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Slot '4a2f8b50-3a1b-4f9e-9d22-123456789abc' is already booked

    get:
      summary: List all bookings
      description: Returns a list of all bookings in the system.
      responses:
        '200':
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Booking'
...
components:
  schemas:
    Slot:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 4a2f8b50-3a1b-4f9e-9d22-123456789abc
        startTime:
          type: string
          format: date-time
          example: '2026-10-05T09:00:00.000Z'
        endTime:
          type: string
          format: date-time
          example: '2026-10-05T09:30:00.000Z'
        isBooked:
          type: boolean
          example: false
        createdAt:
          type: string
          format: date-time
          example: '2026-10-01T08:00:00.000Z'
        updatedAt:
          type: string
          format: date-time
          example: '2026-10-01T08:00:00.000Z'
    CreateBookingRequest:
      type: object
      required:
        - slotId
        - clientName
        - clientEmail
      properties:
        slotId:
          type: string
          format: uuid
          description: UUID of the slot to reserve
          example: 4a2f8b50-3a1b-4f9e-9d22-123456789abc
        clientName:
          type: string
          minLength: 2
          maxLength: 100
          description: Name of the person booking the slot
          example: Ahmed Al-Mansoor
        clientEmail:
          type: string
          format: email
          description: Valid contact email address
          example: ahmed@example.com
    Booking:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 8c3e1a20-5f2c-4e8a-8a11-987654321def
        slotId:
          type: string
          format: uuid
          example: 4a2f8b50-3a1b-4f9e-9d22-123456789abc
        clientName:
          type: string
          example: Ahmed Al-Mansoor
        clientEmail:
          type: string
          format: email
          example: ahmed@example.com
        status:
          type: string
          enum: [CONFIRMED, CANCELLED]
          example: CONFIRMED
        createdAt:
          type: string
          format: date-time
          example: '2026-10-01T10:15:00.000Z'
        updatedAt:
          type: string
          format: date-time
          example: '2026-10-01T10:15:00.000Z'
        slot:
          $ref: '#/components/schemas/Slot'
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Requested resource was not found
    ValidationErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Validation failed
        details:
          type: array
          items:
            type: string
          example:
            - clientEmail is required and must be a valid email address